Patched issues with file share
This code might be buggy. I wrote this update a few days ago, and it works. With this update, I bring file sharing support. There may be exploits due to the location store variable, so if you plan on implementing this code, implement safety measures that prevent certain files from being transmitted, or prevent the other device from storing the file at a given path on the device